fix(HomePage): remove tasks from useEffect deps array #120
this was causing a constant poll to the backend for new todos. The todos
would be refreshed, which would trigger the hook, which trigger new todos
which would trigger the hook, ad infinitum.

See my comment warning about this:
https://github.com/hyper63/hyper-todo-list-demo/pull/18#discussion_r808186208

We just want to initially fetch the todos on the first render, then let
subsequent actions update Redux state, then todos through props triggers
the rerender.