fix(EditPage): minimal changes to get EditPage working #120
the more kosher thing to do here would be to clear out the todo,
but I didn't want to write another reducer and action. So this just uses
a key on the component to force a re-render.

Also properly controls the checkbox on the EditView
See: `isChecked` in https://chakra-ui.com/docs/components/form/checkbox#checkbox-props